Toggling Search Screens (#35)
* search toggle part 1

* toggled the search screens

* adding files to the pr request about toggling search screens

* fix files

* fix files

---------
