+++ package.json
@@ -92,6 +92,7 @@
     "eslint-plugin-expo": "^0.1.0",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^5.0.0",
     "glob": "^11.0.0",
     "husky": "^9.1.5",
