+++ src/navigation/AppNavigator.tsx
@@ -12,13 +12,11 @@ import {
 } from '@/icons';
 import { ContactScreen } from '@/screens/Contact/Contact';
 import { DirectoryScreen } from '@/screens/Directory/Directory';
-import { HomeScreen } from '@/screens/Home/Home';
 import { LoginScreen } from '@/screens/Login/Login';
 import { ShrubSpeciesInfoScreen } from '@/screens/ShrubSpeciesInfo/ShrubSpeciesInfo';
-import { ShrubSpeciesSearchScreen } from '@/screens/ShrubSpeciesSearch/ShrubSpeciesSearch';
+import { SpeciesSearchScreen } from '@/screens/SpeciesSearch/SpeciesSearch';
 import { TreeInfoScreen } from '@/screens/TreeInfo/TreeInfo';
 import { TreeSpeciesInfoScreen } from '@/screens/TreeSpeciesInfo/TreeSpeciesInfo';
-import { TreeSpeciesSearchScreen } from '@/screens/TreeSpeciesSearch/TreeSpeciesSearch';
 import {
   BottomTabParamList,
   ContactStackParamList,
@@ -50,18 +48,10 @@ const LoginStackNavigator = () => {
 const HomeStackNavigator = () => {
   return (
     <HomeStack.Navigator
-      initialRouteName="Home"
+      initialRouteName="SpeciesSearch"
       screenOptions={{ headerShown: false }}
     >
-      <HomeStack.Screen name="Home" component={HomeScreen} />
-      <HomeStack.Screen
-        name="TreeSpeciesSearch"
-        component={TreeSpeciesSearchScreen}
-      />
-      <HomeStack.Screen
-        name="ShrubSpeciesSearch"
-        component={ShrubSpeciesSearchScreen}
-      />
+      <HomeStack.Screen name="SpeciesSearch" component={SpeciesSearchScreen} />
       <HomeStack.Screen
         name="QRCodeScanner"
         component={QRCodeScanner}
