--- src/screens/Home/styles.ts
@@ -1,24 +1,0 @@
-import { StyleSheet } from 'react-native';
-import { colors } from '@/styles/colors';
-import { typography } from '@/styles/typography';
-
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    flexDirection: 'column',
-    justifyContent: 'center',
-  },
-
-  button: {
-    backgroundColor: colors.primary,
-    padding: 16,
-    borderRadius: 6,
-    alignItems: 'center',
-    marginBottom: 12,
-  },
-
-  buttonText: {
-    ...typography.largeBold,
-    color: colors.white,
-  },
-});
