+++ src/screens/Login/Login.tsx
@@ -24,7 +24,7 @@ export const LoginScreen: React.FC<LoginScreenProps> = ({
     setHasLaunched(true);
     navigation.navigate('BottomTabs', {
       screen: 'HomeTab',
-      params: { screen: 'TreeSpeciesSearch' },
+      params: { screen: 'SpeciesSearch' },
     });
   };
