+++ src/types/navigation.ts
@@ -6,8 +6,7 @@ export type LoginStackParamList = {
 
 export type HomeStackParamList = {
   Home: undefined;
-  TreeSpeciesSearch: undefined;
-  ShrubSpeciesSearch: undefined;
+  SpeciesSearch: undefined;
   QRCodeScanner: undefined;
   TreeInfo: { treeId: string };
   TreeSpeciesInfo: { speciesName: string };
